create GithubReleaseExtensions
